main for testing ensemble
There are a couple of tweaks to this main. bilstm MAX_LEN is set to the longest sentence in the test set so that when tested every word has a prediction. Also, some other variables are changed instead of reusing names since they are needed again at the end where the ensemble is scored.